Refactored subscribe_to sql. Added subscribe_from sql.
