+++ blockchain/db/postgres/subscribe_from_db.py
@@ -28,43 +28,3 @@ __license__ = "Apache"
 __version__ = "2.0"
 __maintainer__ = "Joe Roets"
 __email__ = "joe@dragonchain.org"
-
-
-import psycopg2
-import psycopg2.extras
-
-from blockchain.qry import format_node
-
-from postgres import get_connection_pool
-import uuid
-
-""" CONSTANTS """
-DEFAULT_PAGE_SIZE = 1000
-""" SQL Queries """
-SQL_GET_ALL = """SELECT * FROM subscriptions"""
-SQL_INSERT = """INSERT into subscriptions (
-                    sub_id,
-                    subscribe_to,
-                    host,
-                    port,
-                    criteria
-                ) VALUES (%s, %s, %s, %s, %s) """
-
-
-def insert_subscription(subscription):
-    """ insert given subscription into database """
-    values = (
-        str(uuid.uuid4()),
-        subscription['sub_to'],
-        subscription['host'],
-        subscription['port'],
-        psycopg2.extras.Json(subscription['criteria'])
-    )
-    conn = get_connection_pool().getconn()
-    try:
-        cur = conn.cursor()
-        cur.execute(SQL_INSERT, values)
-        conn.commit()
-        cur.close()
-    finally:
-        get_connection_pool().putconn(conn)
